+++ frontend/src/plugins/demo/style.css
@@ -1,13 +1,23 @@
+.bb-guide-cover-wrapper {
+  @apply fixed inset-0;
+  background-color: rgb(33 33 33 / 30%);
+  z-index: 999998;
+}
+
+.bb-guide-target-element {
+  z-index: 999999;
+}
+
 .bb-guide-highlight-wrapper {
   @apply absolute rounded pointer-events-none;
   box-shadow: rgb(33 33 33 / 80%) 0px 0px 1px 2px;
-  z-index: 9999999;
+  z-index: 999999;
 }
 
 .bb-guide-dialog {
   @apply absolute bg-white w-72 m-1 text-black p-2 px-3 rounded-lg text-left;
   box-shadow: 0 3px 30px rgb(33 33 33 / 30%);
-  z-index: 10000000;
+  z-index: 1000000;
 }
 
 .bb-guide-title-text {
